% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cover.example.df}
\alias{Cover.example.df}
\title{Cover Dataset}
\format{
A data frame with 10 columns:
\describe{
  \item{x}{X coordinate of the study site}
  \item{y}{Y coordinate of the study site}
  \item{\emph{Leaves}}{Presence (1) or absence (0) of leaves as cover in each pixel/microhabitat }
  \item{\emph{Algae}}{Presence (1) or absence (0) of algae as cover in each pixel/microhabitat }
  \item{\emph{Root}}{Presence (1) or absence (0) of roots as cover in each pixel/microhabitat }
  \item{\emph{Aquatic.vegetation}}{Presence (1) or absence (0) of aquatic vegetation in each pixel/microhabitat }
  \item{\emph{Reed}}{Presence (1) or absence (0) of reeds and emergent vegetation in each pixel/microhabitat }
  \item{\emph{Wood}}{Presence (1) or absence (0) of woody debris and logs as cover in each pixel/microhabitat }
  \item{\emph{Sand}}{Presence (1) or absence (0) of sand in each pixel/microhabitat }
  \item{\emph{Rock}}{Presence (1) or absence (0) of rocks in each pixel/microhabitat }
  \item{\emph{Cave}}{Presence (1) or absence (0) of caves cover in each pixel/microhabitat }
  \item{\emph{Shade}}{Presence (1) or absence (0) of shade in each pixel/microhabitat }
}
}
\usage{
Cover.example.df
}
\description{
This dataset contains the presence (1) or absence (0) of various types of cover/shelter in every pixel/microhabitat within the study site in the Serpis River Basin (Spain),
where the hydraulic simulation was carried out.
}
\examples{
data(Cover.example.df)
summary(Cover.example.df)
}
\references{
Mu\ifelse{html}{\out{&ntilde;}}{\enc{UTF-8}{ñ}}oz-Mas, R., Macian-Sorribes, H., Oliva-Paterna, F. J.,
Sangelantoni, L., Peano, D., Pulido-Velazquez, M., & Mart\ifelse{html}{\out{&iacute;}}{\enc{UTF-8}{í}}nez-Capel, F. (2024).
Adaptation measures to global change in the Serpis River Basin (Spain): An evaluation considering agricultural benefits,
environmental flows, and invasive fishes. \emph{Ecological Indicators}, 161, 111979.
DOI: 10.1016/j.ecolind.2024.111979
}
\keyword{datasets}
