% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverIndex.R
\name{Cover.index}
\alias{Cover.index}
\title{Title}
\usage{
Cover.index(
  Covers = NA,
  Leaves = NA,
  Algae = NA,
  Root = NA,
  Aquatic.vegetation = NA,
  Reed = NA,
  Wood = NA,
  Sand = NA,
  Rock = NA,
  Cave = NA,
  Shade = NA,
  Species,
  Size.class,
  model
)
}
\arguments{
\item{Model}{}
}
\description{
Title
}
