% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IberianFishHSMs-package.R
\docType{package}
\name{IberianFishHSMs}
\alias{IberianFishHSMs-package}
\alias{IberianFishHSMs}
\title{Package-level documentation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The modelling approaches include sets of Habitat Suitability Curves (HSCs), Artificial Neural Networks (Multilayer Perceptrons, ANN-MLPs), Random Forests (RFs), Generalised Additive Models (GAMs), and Support Vector Machines (SVMs), all based on empirical data collected across several river basins in the Iberian Peninsula. The HSCs have been reviewed by an expert panel. The package also compiles sets of HSCs and Fuzzy Rule-Based Systems (FRBSs) based on expert knowledge, which was gathered through online forms. Functions for evaluating habitat suitability from hydraulic simulations and for plotting the modelled suitability are also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RafaMMas/IberianFishHSMs}
  \item Report bugs at \url{https://github.com/RafaMMas/IberianFishHSMs/issues}
}

}
\author{
\strong{Maintainer}: Rafael Muñoz-Mas \email{rafa.m.mas@gmail.com} (\href{https://orcid.org/0000-0002-7263-5799}{ORCID})

Other contributors:
\itemize{
  \item Javier Gortázar (\href{https://orcid.org/0009-0005-3028-6921}{ORCID}) [contributor]
  \item Enric Aparicio (\href{https://orcid.org/0000-0001-8810-0783}{ORCID}) [contributor]
  \item Fernando Cobo (\href{https://orcid.org/0000-0001-5684-266X}{ORCID}) [contributor]
  \item Francisco J. Oliva-Paterna (\href{https://orcid.org/0000-0001-8288-5321}{ORCID}) [contributor]
  \item Francisco Martínez-Capel (\href{https://orcid.org/0000-0003-4991-0251}{ORCID}) [contributor]
  \item Gustavo González Fernández [contributor]
  \item José Prenda (\href{https://orcid.org/0000-0002-1505-8331}{ORCID}) [contributor]
  \item Rafael Miranda (\href{https://orcid.org/0000-0003-4798-314X}{ORCID}) [contributor]
}

}
\keyword{internal}
