% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListSelectedCoverTypes.R
\name{ListSelectedCoverTypes}
\alias{ListSelectedCoverTypes}
\title{Title}
\usage{
ListSelectedCoverTypes(
  Species = NULL,
  Size = NULL,
  River = NULL,
  Model.type = NULL,
  Sampled.season = NULL,
  Valid.season = NULL,
  Data.origin = NULL,
  Default = TRUE,
  verbose = TRUE,
  Only.models = FALSE
)
}
\arguments{
\item{Only.models}{}
}
\description{
Title
}
