% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHabitatSuitabilityModels.R
\name{PlotHabitatSuitabilityModels}
\alias{PlotHabitatSuitabilityModels}
\title{Title}
\usage{
PlotHabitatSuitabilityModels(
  Selected.model = NULL,
  data = NULL,
  n.points = 50,
  Quantiles = TRUE,
  Layout = c(2, 2),
  HSC.aggregation = "geometric"
)
}
\arguments{
\item{HSC.aggregation}{}
}
\description{
Title
}
