% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictHabitatSuitability.R
\name{PredictHabitatSuitability}
\alias{PredictHabitatSuitability}
\title{Title}
\usage{
PredictHabitatSuitability(
  Selected.models = NULL,
  data = NULL,
  HSC.aggregation = "geometric"
)
}
\arguments{
\item{HSC.aggregation}{}
}
\description{
Title
}
