% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictNNETensemble.R
\name{PredictNNETensemble}
\alias{PredictNNETensemble}
\title{Predict suitability from SVM ensembles employing mean votes}
\usage{
PredictNNETensemble(object, newdata, type = "raw")
}
\arguments{
\item{object}{Ensemble of Artificial Neural Networks - Multilayer Perceptrons (NNets) as a list}

\item{newdata}{data frame with four columns: Velocity, Depth, Substrate.index and Cover}

\item{type}{string "raw" or "class" determining the kind of output suitability between 0 and 1 or present ("0") or absent ("1")}
}
\value{
predicted suitability or class
}
\description{
Predict suitability from SVM ensembles employing mean votes
}
