% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictHabitatSuitability.R
\name{PredictSuitabilityPlot}
\alias{PredictSuitabilityPlot}
\title{Title}
\usage{
PredictSuitabilityPlot(
  Selected.model = NULL,
  data = NULL,
  HSC.aggregation = "geometric"
)
}
\arguments{
\item{HSC.aggregation}{}
}
\description{
Title
}
