% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictSVMensemble.R
\name{PredictSVMensemble.mean.votes}
\alias{PredictSVMensemble.mean.votes}
\title{Predict suitability from SVM ensembles employing mean votes}
\usage{
PredictSVMensemble.mean.votes(object, newdata)
}
\arguments{
\item{object}{Ensemble of Support Vector Machines (SVMs) as a list}

\item{newdata}{data frame with four columns: Velocity, Depth, Substrate.index and Cover}
}
\value{
predicted suitability
}
\description{
Predict suitability from SVM ensembles employing mean votes
}
