% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Substrate.index.example.df}
\alias{Substrate.index.example.df}
\title{Substrate Index Dataset}
\format{

}
\usage{
Substrate.index.example.df
}
\description{
This dataset contains granulometry data from the study site where the hydraulic simulation was carried out in the Serpis River Basin (Spain).
The columns represent the percentage of various substrate types, including different types of sediments and bedrock.
}
\examples{
data(Substrate.index.example.df)
summary(Substrate.index.example.df)
}
\references{
Mu\ifelse{html}{\out{&ntilde;}}{\enc{UTF-8}{ñ}}oz-Mas, R., Macian-Sorribes, H., Oliva-Paterna, F. J.,
Sangelantoni, L., Peano, D., Pulido-Velazquez, M., & Mart\ifelse{html}{\out{&iacute;}}{\enc{UTF-8}{í}}nez-Capel, F. (2024).
Adaptation measures to global change in the Serpis River Basin (Spain): An evaluation considering agricultural benefits,
environmental flows, and invasive fishes. \emph{Ecological Indicators}, 161, 111979.
DOI: 10.1016/j.ecolind.2024.111979
}
\keyword{datasets}
