% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubstrateIndex.R
\name{SubstrateIndex}
\alias{SubstrateIndex}
\title{Calculate Substrate Index}
\usage{
SubstrateIndex(Substrate, check.completeness = TRUE)
}
\arguments{
\item{Substrate}{A data frame containing at least the following columns:
"Sand", "Fine.gravel", "Gravel", "Cobbles", "Boulders", and "Bed.rock". Column names must coincide.
Values should be between 0 and 100, representing the percentage of each class.}

\item{check.completeness}{Logical, if TRUE (default), checks whether the percentages in each pixel/microhabitat sum up to 100.}
}
\value{
A numeric index ranging from 0 (fine substrate, e.g., silt) to 8 (bedrock).
}
\description{
Computes the substrate index as a weighted aggregation of the percentage
of different granulometry classes (%).
}
\details{
The function aggregates the substrate cover percentages using a weighted formula
  to derive an index representing substrate composition, where higher values correspond
  to coarser substrates.
}
\examples{
data(Substrate.index.example.df)
summary(Substrate.index.example.df)
\dontrun{
SubstrateIndex(Substrate.index.example.df)
}
SubstrateIndex(Substrate.index.example.df, check.completeness = FALSE)

}
