% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Velocity.example.df}
\alias{Velocity.example.df}
\title{Hydraulic Simulation Velocity Dataset}
\format{
A data frame with 41 columns:
\describe{
\item{x}{X coordinate in the hydraulic simulation}
\item{y}{Y coordinate in the hydraulic simulation}
\item{Velocity.0.05}{Simulation with 0.05 cubic meters per second (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}}/s), velocity values in meters (m/s)}
\item{Velocity.0.1}{Simulation with 0.1 cubic meters per second (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}}/s), velocity values in meters (m/s)}
\item{Velocity.0.2}{Simulation with 0.2 cubic meters per second (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}}/s), velocity values in meters (m/s)}
\item{...}{Additional velocity columns for increasing flow scenarios up to 21 cubic meters per second (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}}/s)}
\item{Velocity.21}{Simulation with 21 cubic meters per second (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}}/s), velocity values in meters (m/s)}
}
}
\usage{
Velocity.example.df
}
\description{
This dataset contains siulated velocity from a hydraulic model. The data was obtained from a study on the Serpis River Basin (Spain),
assessing adaptation measures to global change.
}
\examples{
data(Velocity.example.df)
summary(Velocity.example.df)
}
\references{
Mu\ifelse{html}{\out{&ntilde;}}{\enc{UTF-8}{ñ}}oz-Mas, R., Macian-Sorribes, H., Oliva-Paterna, F. J.,
Sangelantoni, L., Peano, D., Pulido-Velazquez, M., & Mart\ifelse{html}{\out{&iacute;}}{\enc{UTF-8}{í}}nez-Capel, F. (2024).
Adaptation measures to global change in the Serpis River Basin (Spain): An evaluation considering agricultural benefits,
environmental flows, and invasive fishes. \emph{Ecological Indicators}, 161, 111979.
DOI: 10.1016/j.ecolind.2024.111979
}
\keyword{datasets}
