% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListModels.R
\name{ListModels}
\alias{ListModels}
\title{List available models and their key characteristics.}
\usage{
ListModels(
  Species = NULL,
  Size = NULL,
  River = NULL,
  Model.type = NULL,
  Sampled.season = NULL,
  Valid.season = NULL,
  Data.origin = NULL,
  Default = TRUE,
  verbose = TRUE,
  Only.models = FALSE
)
}
\arguments{
\item{Species}{A string specifying the selected species (e.g., "Salmo trutta").}

\item{Size}{A string specifying the selected size class (e.g., "Very small", "Small", "Medium", "Large", "Spawning").}

\item{River}{A string specifying the selected river. Since rivers vary, it is recommended to first filter by species before selecting a specific river.}

\item{Model.type}{A string specifying the selected model type (i.e., "FRBS", "GAM", "HSC", "NNET", "RF", "SVM").
\describe{
  \item{FRBS: Fuzzy Rule-Based System.}
  \item{GAM: Generalized Additive Model.}
  \item{HSC: Habitat Suitability Curve/Criteria.}
  \item{NNET: Artificial Neural Network (Multilayer Perceptron).}
  \item{RF: Random Forest.}
  \item{SVM: Support Vector Machine.}
}}

\item{Sampled.season}{The sampling period during which data was collected on one or multiple occasions.}

\item{Valid.season}{The time period during which the developed models remain applicable before requiring new models.}

\item{Data.origin}{The source of the data used to develop the habitat suitability model.}

\item{Default}{Logical (`TRUE`/`FALSE`). If `TRUE`, only the recommended models are listed.}

\item{verbose}{Logical. If `TRUE`, the table of available models is printed.}

\item{Only.models}{Logical. If `TRUE`, the output is limited to the model codes, which are internally used to call the corresponding habitat suitability model.}
}
\value{
A list containing the characteristics of the selected models. The structure of the list depends on the value of
\code{Only.models}:
\describe{
  \item{If \code{Only.models = FALSE} (default):}{The function returns a list with three elements:
    \describe{
      \item{\code{$Current.summary.table}}{A data frame containing the available models and their characteristics.}
      \item{\code{$Models}}{A character vector containing the names of the models along with a summary of their main characteristics.}
      \item{\code{$Codes}}{A character vector of model codes, which are internally used to call the corresponding habitat suitability model.}
    }
  }
  \item{If \code{Only.models = TRUE}:}{The function returns a list with a single element:
    \describe{
      \item{\code{$Codes}}{A character vector of model codes, which are internally used to call the corresponding habitat suitability model.}
    }
  }
}
}
\description{
The queries can be categorized by Species, Size, River, Model.type, Sampled.season, Valid.season, and/or Data.origin.
}
\examples{

ListModels(Species = "Salmo trutta")
ListModels(Species = "Salmo trutta", Only.models = TRUE)
ListModels(Species = "Salmo trutta", verbose = FALSE, Only.models = TRUE)

}
