% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictHabitatSuitability.R
\name{Load.and.print.model}
\alias{Load.and.print.model}
\title{Load and prints the selected model}
\usage{
Load.and.print.model(Selected.model = NULL)
}
\arguments{
\item{Selected.model}{A character string or object representing the selected habitat suitability model. It should be among the \code{Codes} obtained with \code{ListModels} Default is \code{NULL}.}
}
\value{
List containing the selected cover types (\code{Selected.cover.types}), the model type (\code{Model.type}) and the model itself (\code{Model}).
}
\description{
Load and prints the selected model
}
\examples{

(Selected.model <- ListModels(Species = "Salmo trutta", verbose = FALSE)$Codes[1])

Load.and.print.model(Selected.model = Selected.model)

}
