% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHabitatSuitabilityModels.R
\name{PredictSuitabilityPlot}
\alias{PredictSuitabilityPlot}
\title{Function to determine the microhabitat suitability within PlotHabitatSuitabilityModels}
\usage{
PredictSuitabilityPlot(
  Selected.model = NULL,
  data = NULL,
  HSC.aggregation = "geometric"
)
}
\arguments{
\item{Selected.model}{A character string or object representing the selected habitat suitability model. It should be among the \code{Codes} obtained with \code{ListModels} Default is \code{NULL}.}

\item{data}{A data frame containing the data to be used for plotting the habitat suitability models. Default is \code{NULL}. This dataset must contain: "Velocity", "Depth", "Substrate.index" and "Cover".}

\item{HSC.aggregation}{A string specifying the aggregation method for habitat suitability model data. Options are \code{"prod"}, \code{"min"}, \code{"geometric"}, \code{"arithmetic"}. Default is \code{"geometric"}.}
}
\value{
suitability betwen 0 and 1.
}
\description{
Function to determine the microhabitat suitability within PlotHabitatSuitabilityModels
}
\details{
This function is internally called.
}
