% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictSVMensemble.R
\name{PredictSVMensemble.probability}
\alias{PredictSVMensemble.probability}
\title{Title}
\usage{
PredictSVMensemble.probability(object, newdata)
}
\arguments{
\item{newdata}{}
}
\description{
Title
}
