% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubstrateIndex.R
\name{SubstrateIndex}
\alias{SubstrateIndex}
\title{Title}
\usage{
SubstrateIndex(
  Substrate = NULL,
  Vegetation = NULL,
  Silt = NULL,
  Sand = NULL,
  Fine.gravel = NULL,
  Gravel = NULL,
  Cobbles = NULL,
  Boulders = NULL,
  Bed.rock = NULL,
  check.completeness = TRUE
)
}
\arguments{
\item{check.completeness}{}
}
\description{
Title
}
